Move libraries used by server into deps
Move logger.js into /server.
Remove react-tooltip module.
Remove .flowconfig file, which is not needed.
