--- .flowconfig
@@ -1,11 +1,0 @@
-[ignore]
-.*/build
-.*/docs
-.*/node_modules
-.*/public
-
-[include]
-
-[options]
-module.system.node.resolve_dirname=node_modules
-module.system.node.resolve_dirname=src
